pyproject.toml
[tool.poetry]
name = "pythonic news api"
version = "0.1.0"
description = "API of articles"
authors = ["MatheusLeall <matheuscardoso.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
djangorestframework = "^3.14.0"
requests = "^2.28.2"
newsapi-python = "^0.2.7"
pytest = "^7.3.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
